break out components into folders
